move various 'help' related handlers to Help.java; make non-dynamic where appropriate
